pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "enjoy-python"
version = "0.1.0"
description = "A Python project for automating blog interactions"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "selenium",
    "google-generativeai",
    "requests",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-mock",
    "pytest-asyncio"
]

[tool.setuptools.packages.find]
where = ["src"]
